refactor: share markdown sync manifest loader
